+++ Helper/Data.php
@@ -7,6 +7,7 @@
 
 namespace MageRocket\GoCuotas\Helper;
 
+use Magento\Sales\Model\Order;
 use MageRocket\GoCuotas\Logger\Logger;
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\Exception\NoSuchEntityException;
@@ -405,18 +406,50 @@ class Data
     }
 
     /**
+     * Get CronOrderTimeout
+     *
+     * @param int|null $storeId
+     * @return string
+     */
+    public function getCronOrderTimeout($storeId = null): string
+    {
+        return $this->getConfig('cron/order_timeout', $storeId) ?? 30;
+    }
+
+    /**
      * Generate Token
      *
-     * @param int $order
+     * @param Order $order
      * @param string|null $status
      * @return string
      */
-    public function generateToken($order, $status = null): string
+    public function generateToken(Order $order, $status = null): string
     {
         return hash(
             'sha256',
             $this->getSecret($order->getStoreId()) . $order->getIncrementId() . $order->getCreatedAt() . $status
         );
+    }
+
+    /**
+     * maskSensitiveData
+     *
+     * @param $data
+     * @param int $visibleStart
+     * @param int $visibleEnd
+     * @param string $maskChar
+     * @return string
+     */
+    function maskSensitiveData($data, int $visibleStart = 3, int $visibleEnd = 2, string $maskChar = '*'): string
+    {
+        $length = strlen($data);
+        if ($length <= ($visibleStart + $visibleEnd)) {
+            return $data;
+        }
+        $start = substr($data, 0, $visibleStart);
+        $end = substr($data, -$visibleEnd);
+        $masked = str_repeat($maskChar, $length - ($visibleStart + $visibleEnd));
+        return $start . $masked . $end;
     }
 
     /**