+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.4] - 2024-11-25
+
+### Added
+- New configuration added that allows the merchant to modify the waiting time before canceling an order.
+
+### Updated
+- Cron modified to check with GoCuotas if there is an unreported payment for the order before canceling it.
+
 ## [1.0.2/3] - 2024-07-08
 
 ### Added